chore(discord-bot): fix circular deps
