+++ discord-bot/src/Client/index.ts
@@ -1,8 +1,8 @@
+import { config, error, info } from '#'
 import { commands } from 'commands'
 import { DiscordClient } from 'DiscordClient'
 import { Service } from 'typedi'
-import { config, error, info } from '~'
-import { Events } from '../events'
+import { Events } from '../events/Events'
 import { ClientI } from './interface'
 
 @Service()
