+++ discord-bot/src/DiscordClient.ts
@@ -1,6 +1,6 @@
+import { config } from '#/config'
 import { Client } from 'discord.js'
 import { Service } from 'typedi'
-import { config } from '~'
 
 @Service()
 export class DiscordClient extends Client {
