+++ discord-bot/src/commands/VerifyCommand.ts
@@ -1,5 +1,5 @@
+import { config } from '#/config'
 import { ChannelType, CommandInteraction, PermissionsBitField, SlashCommandBuilder } from 'discord.js'
-import { config } from '~'
 import { _Command } from './_Command'
 import { CommandName } from './interface'
