+++ discord-bot/src/lib/verify-onchain.ts
@@ -1,8 +1,8 @@
 // TODO: make a package of https://github.com/privacy-scaling-explorations/e2e-zk-ecdsa/blob/main/circuits/circom/scripts/verify-onchain.ts
 
 import { groth16 } from 'snarkjs'
-import { config } from '~'
 import { groth16VerifierAbi as abi } from './abis/groth16Verifier'
+import { config } from './config'
 import { Proof, PublicSignals } from './types'
 import { web3Client } from './web3-client'
